Fixed normalizePayload, extractMeta & extractLinks for json-api url shorthand style.
The current versions of the DS.RESTAdapter and DS.RESTSerializer call normalizePayload with only 1 parameter. I removed `type` parameter to fix this.

I am trying to implement this http://jsonapi.org/format/#document-url-template-shorthands, but I didn't try if this change breaks another of the JSON API formats.
Unfortunately there are no tests for this project at the moment.